Don't start a core in CoreApplication's dtor... fixes BR #322

+++ src/core/coreapplication.cpp
@@ -22,7 +22,9 @@
 
 #include "core.h"
 
-CoreApplicationInternal::CoreApplicationInternal() {
+CoreApplicationInternal::CoreApplicationInternal() 
+  : _coreCreated(false)
+{
   // put core-only arguments here
   CliParser *parser = Quassel::cliParser();
   parser->addOption("port",'p', tr("The port quasselcore will listen at"), QString("4242"));
@@ -33,8 +35,10 @@ CoreApplicationInternal::CoreApplicationInternal() {
 }
 
 CoreApplicationInternal::~CoreApplicationInternal() {
-  Core::saveState();
-  Core::destroy();
+  if(_coreCreated) {
+    Core::saveState();
+    Core::destroy();
+  }
 }
 
 bool CoreApplicationInternal::init() {
@@ -51,6 +55,7 @@ bool CoreApplicationInternal::init() {
   }
 
   Core::instance();  // create and init the core
+  _coreCreated = true;
 
   if(!Quassel::isOptionSet("norestore")) {
     Core::restoreState();
